- Moved save_posterior_model function and related helpers from fitting to psfMC.analysis.images. - Major refactoring of plotting functions. - plotchain script now uses difflib.SequenceMatcher to find the closest-matching model file name in the directory for each supplied database. - model parser now changes to the model file directory before compiling the model file (and changes back afterward). This ensures that all relative path references in the model file are relative to the model file location, instead of the calling directory.
