+++ examples/run_example.py
@@ -1,11 +1,7 @@
 from psfMC import model_galaxy_mcmc
+from psfMC.analysis import plot_hist
 import glob
 import subprocess
-import pymc
-try:
-    import matplotlib.pyplot as pp
-except ImportError:
-    pp = None
 
 # Arguments that will be passed to the MCMC sampler. iter and burn are selected
 # to allow this demonstration to run quickly. They are almost certainly too
@@ -25,15 +21,10 @@ for model_file in model_files:
     model_galaxy_mcmc(model_file, output_name=output_name, **mc_args)
 
     # Once sampling has completed, display some example plots
-    if pp is not None:
-        db_module = getattr(pymc.database, mc_args['backend'])
-        db = db_module.load(output_name + '_db.' + mc_args['backend'])
-        for trace_name in ('0_Sky_adu', '1_PSF_mag', '2_Sersic_mag',
-                           '2_Sersic_reff', '2_Sersic_index', '2_Sersic_angle'):
-            pp.hist(db.trace(trace_name)[:], bins=20)
-            pp.title(trace_name)
-            pp.show()
-        db.close()
+    db_file = output_name + '_db.' + mc_args['backend']
+    for trace_name in ('0_Sky_adu', '1_PSF_mag', '1_PSF_2_Sersic_magdiff'):
+        plot_hist(trace_name, db_file, model=model_file,
+                  chains=range(-mc_args['chains'], 0))
 
     # Also try to open a ds9 window with all the input and output fits images
     in_images = glob.glob('sci_*.fits')
