+++ psfMC/__init__.py
@@ -1,5 +1,6 @@
 __version__ = 'v1.0b4'
 
-from .fitting import model_galaxy_mcmc, save_posterior_model
+from .fitting import model_galaxy_mcmc
 from .models import multicomponent_model
-from .analysis import potential_scale_reduction, num_effective_samples
+from .analysis import save_posterior_images, potential_scale_reduction, \
+    num_effective_samples
