+++ psfMC/analysis/__init__.py
@@ -1,3 +1,4 @@
 from .statistics import potential_scale_reduction, num_effective_samples, \
     max_posterior_sample, calculate_dic, check_convergence_psrf, chain_autocorr
 from .plotting import plot_hist, plot_trace
+from .images import save_posterior_images
